+++ tests/testthat/test-aaa-deprecated.R
@@ -37,3 +37,18 @@ test_that("on", {
   expect_identical(chk_off(), list(chk.on = FALSE))
 })
 
+test_that("vld_no_missing", {
+  expect_true(vld_no_missing(1))
+  expect_true(vld_no_missing(integer(0)))
+  expect_false(vld_no_missing(NA))
+  expect_false(vld_no_missing(c(NA, 1)))
+})
+
+test_that("chk_no_missing", {
+  expect_null(chk_no_missing(1))
+  expect_invisible(chk_no_missing(1))
+  expect_chk_error(chk_no_missing(NA), "^`NA` must not have missing values[.]$")
+  expect_chk_error(chk_no_missing(NA, x_name = "1"), "^1 must not have missing values[.]$")
+})
+
+